tests/comprehensive/test_comprehensive.py
#!/usr/bin/env python3
import asyncio
import json
import os
import tempfile
import unittest
from unittest.mock import Mock, patch, MagicMock
import sys

# Add framework to path
# Framework imports (updated for domain-driven structure)
from framework import App
from framework.core import Component, Config, Container, Scope
from framework.data.db import SQLAlchemyAdapter
from framework.web import HttpServer
from framework.core.plugins import PluginManager
from framework.monitoring import Logging
from framework.monitoring.metrics import HTTPMetricsAdapter
from framework.monitoring import Observability
from framework.processing.cli import cli
from framework.messaging.redis_adapter import RedisAdapter
from framework.simulation import EnvironmentEntity
from framework.simulation.controllers import Transform

class TestFrameworkComprehensive(unittest.IsolatedAsyncioTestCase):
    """Comprehensive test suite for UCore Framework with maximum coverage."""

    def setUp(self):
        """Set up test environment before each test."""
        self.test_dir = tempfile.mkdtemp()
        self.config_file = os.path.join(self.test_dir, 'test_config.json')
        self.db_file = os.path.join(self.test_dir, 'test.db')
        self.log_file = os.path.join(self.test_dir, 'test.log')

        import yaml
        with open(self.config_file, 'w') as f:
            yaml.dump({
                "app": {"name": "TestApp", "version": "1.0.0"},
                "database": {"url": f"sqlite:///{self.db_file}"},
                "logging": {"level": "DEBUG", "file": self.log_file}
            }, f)

    def tearDown(self):
        """Clean up test environment after each test."""
        for file in [self.config_file, self.db_file, self.log_file]:
            if os.path.exists(file):
                os.unlink(file)
        if os.path.exists(self.test_dir):
            os.rmdir(self.test_dir)

    @patch('argparse.ArgumentParser')
    async def test_app_initialization_comprehensive(self, mock_arg_parser):
        """Test comprehensive app initialization scenarios."""
        mock_args = MagicMock()
        mock_args.config = self.config_file
        mock_args.log_level = "DEBUG"
        mock_args.plugins_dir = None
        mock_arg_parser.return_value.parse_args.return_value = mock_args

        app = App(name="TestApp")
        app.bootstrap(mock_args)

        self.assertEqual(app.name, "TestApp")
        self.assertIsNotNone(app.logger)
        self.assertIsNotNone(app.container)

        class LifecycleComponent(Component):
            def __init__(self, app: App):
                super().__init__(app)
                self.started = False
                self.stopped = False
            async def start(self):
                self.started = True
            async def stop(self):
                self.stopped = True

        app.register_component(LifecycleComponent)
        component = app._components[0]

        await app.start()
        self.assertTrue(component.started)
        await app.stop()
        self.assertTrue(component.stopped)

    async def test_component_registration_and_lifecycle(self):
        """Test component registration with lifecycle management."""
        app = App(name="TestApp")

        class MockDB(Component):
            def __init__(self, app: App):
                super().__init__(app)
                self.started = False
                self.stopped = False
            async def start(self):
                await asyncio.sleep(0.01)
                self.started = True
            async def stop(self):
                await asyncio.sleep(0.01)
                self.stopped = True

        class MockCache(Component):
            def __init__(self, app: App):
                super().__init__(app)

        app.register_component(MockDB)
        app.register_component(MockCache)

        self.assertEqual(len(app._components), 2)
        db_comp = app._components[0]

        await app.start()
        self.assertTrue(db_comp.started)
        await app.stop()
        self.assertTrue(db_comp.stopped)

    def test_config_comprehensive(self):
        """Test configuration management comprehensively."""
        config = Config()
        config.load_from_file(self.config_file)
        self.assertEqual(config.get('app.name'), 'TestApp')
        os.environ['UCORE_APP_VERSION'] = '2.0.0'
        config.load_from_env()
        self.assertEqual(config.get('app.version'), '2.0.0')
        del os.environ['UCORE_APP_VERSION']

    def test_dependency_injection_comprehensive(self):
        """Test dependency injection container comprehensively."""
        container = Container()

        class DatabaseService:
            pass
        class UserService:
            def __init__(self, db: DatabaseService):
                self.db = db

        container.register(DatabaseService, scope=Scope.SINGLETON)
        container.register(UserService, scope=Scope.TRANSIENT)

        db1 = container.get(DatabaseService)
        db2 = container.get(DatabaseService)
        self.assertIs(db1, db2)

        user1 = container.get(UserService)
        user2 = container.get(UserService)
        self.assertIsNot(user1, user2)
        self.assertIs(user1.db, user2.db)

    @patch('sqlalchemy.create_engine')
    def test_sqlalchemy_adapter_comprehensive(self, mock_create_engine):
        """Test SQLAlchemy adapter comprehensively."""
        app = App("TestDBApp")
        config = app.container.get(Config)
        config.set("database.url", f"sqlite:///{self.db_file}")
        
        db_adapter = SQLAlchemyAdapter(app)
        db_adapter.connect()
        mock_create_engine.assert_called_with(f"sqlite:///{self.db_file}")

    async def test_http_server_comprehensive(self):
        """Test HTTP server functionality comprehensively."""
        app = App("TestHTTPApp")
        http_server = HttpServer(app)
        # Note: HttpServer already registers itself in the container during __init__

        @http_server.route("/test", "GET")
        async def test_handler(request):
            return {"message": "test response"}

        # Check that route was added to the web app router
        self.assertTrue(len(http_server.web_app.router.routes()) > 0)

    @patch('os.makedirs')
    @patch('builtins.open')
    def test_logging_comprehensive(self, mock_open, mock_makedirs):
        """Test logging system comprehensively."""
        logging = Logging()
        logger = logging.get_logger("test_module")
        self.assertIsNotNone(logger)
        self.assertEqual(logger.name, "test_module")

    @patch('framework.monitoring.observability.requests.get')
    def test_observability_comprehensive(self, mock_get):
        """Test observability features comprehensively."""
        mock_response = Mock()
        mock_response.status_code = 200
        mock_get.return_value = mock_response
        
        observability = Observability()
        result = observability.probe_endpoint("http://example.com/health")
        self.assertTrue(result)

    def test_metrics_comprehensive(self):
        """Test metrics collection comprehensively."""
        app = App("TestMetricsApp")
        metrics = HTTPMetricsAdapter(app)
        # Test that HttpServer can be instantiated without conflicts
        http_instance = HttpServer(app)
        # This is a simplified test. A full test would involve running the http server.
        self.assertIsNotNone(metrics.http_requests_total)

    @patch('typer.main.get_command')
    def test_cli_comprehensive(self, mock_get_command):
        """Test CLI interface comprehensively."""
        # This is a simplified test. A full test would involve running the CLI.
        self.assertIsNotNone(cli)

    @patch('aioredis.from_url')
    async def test_redis_adapter_comprehensive(self, mock_redis):
        """Test Redis adapter comprehensively."""
        # Skip if aioredis is not available (it would be in full installation)
        try:
            import aioredis
        except ImportError:
            self.skipTest("aioredis not installed - skipping Redis integration test")

        mock_client = MagicMock()
        mock_redis.return_value = mock_client
        mock_client.get.return_value = "test_value"

        app = App("TestApp")
        adapter = RedisAdapter(app)
        adapter.redis = mock_client

        value = await adapter.get("test_key")
        self.assertEqual(value, "test_value")
        await adapter.set("test_key", "new_value")
        mock_client.set.assert_called_with("test_key", "new_value")

    def test_simulation_comprehensive(self):
        """Test simulation framework comprehensively."""
        entity = EnvironmentEntity(name="TestEntity")
        transform = Transform(x=10, y=20)
        entity.add_controller(transform)
        self.assertEqual(transform.x, 10)

if __name__ == '__main__':
    unittest.main()
