requirements.txt
# UCore Framework Requirements
# Production-ready, enterprise-grade framework dependencies

# Core Framework
aiohttp>=3.9.0          # Async HTTP server
pydantic>=2.5.0         # Data validation
typing-extensions>=4.8.0  # Enhanced type hints

# Message Bus & Data
redis>=5.0.0            # Redis client for message bus
celery>=5.3.0           # Background task processing

# CLI & User Experience
typer>=0.12.0           # CLI framework
rich>=13.7.0           # Beautiful terminal formatting
click>=8.1.0           # CLI foundation

# Testing & Quality
pytest>=7.4.0          # Testing framework
pytest-asyncio>=0.21.0 # Async testing
pytest-mock>=3.11.0    # Mock support for testing
coverage>=7.3.0       # Test coverage
flake8>=6.1.0         # Code linting

# Development & Documentation
black>=23.10.0         # Code formatting
mypy>=1.6.0           # Type checking
sphinx>=7.2.0         # Documentation generation
mkdocs>=1.5.0         # Static site documentation
mkdocs-material>=9.4.0 # Material theme for docs

# UI Frameworks (Optional)
PySide6>=6.6.0        # Desktop applications
flet>=0.20.0          # Web UIs

# Database (Optional - for full database integration)
sqlalchemy>=2.0.0     # ORM
aiosqlite>=0.19.0     # Async SQLite driver
alembic>=1.12.0       # Database migrations

# Observability (Optional)
prometheus-client>=0.17.0  # Metrics collection

# Configuration & Logging
pyyaml>=6.0.0         # YAML configuration
loguru>=0.7.0         # Enhanced logging
python-dotenv>=1.0.0  # Environment variables

# Screen Capture (for ScreenCapturer controller)
mss>=9.0.0           # Fast screen capture library
pyautogui>=0.9.50     # Cross-platform GUI automation
Pillow>=9.0.0         # Image processing for screenshots

# DuckDuckGo Search (for image search examples)
duckduckgo-search>=6.1.0  # DuckDuckGo search API

# Game Engine (for Render controller)
pygame>=2.5.0        # 2D game engine for simulation visualization
